Add new service to retrieve a summary of all lists.
This route only returns an array of objects (lists and watchlist). Each list only has 3 properties: name, id, movies. Movies is an array that includes the idTMDBs of the list.

This route will allow to implement the addition of movies to lists later.

Add test and the necessary change to the files for testing.